mclient/requirements.txt
docker==6.1.2
websockets==11.0.3
pydantic==2.1.1
pydantic-settings==2.0.3
python-dotenv==1.0.0
psutil==5.9.5
aiohttp==3.9.5
prompt_toolkit==3.0.39